requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
markdown>=3.5.0
geopy>=2.4.0
folium>=0.15.0
pandas>=2.1.0
tabulate>=0.9.0
python-dotenv>=1.0.0
PyQt6>=6.6.0
PyQt6-WebEngine>=6.6.0
pyinstaller>=6.0.0
